base: Make use of the brightness slider buttons
Introduced in the commit below were two buttons that were used to mirror
the Oxygen OS look. They did absolutely nothing until now.

https://github.com/DirtyUnicorns/android_frameworks_base/commit/7f567fde01da0d313d123d711afa915cb05e4ee8

This works in two ways, single clicks and long clicks

Single click = decreases and increases the brightness in intervals of 10
Long click = sets the brightness to low or high depending on which side you pick

- Check for auto brightness mode and disable it if enabled before performing actions

Thanks to xdnax, ezio84 and Thecrazyskull for the idea/suggestions

Change-Id: I3267439145a9d2bbc16cb6c55586a7403472381c
Signed-off-by: spezi77 <spezi7713@gmx.net>
Signed-off-by: Aston-Martinn <advaithbhat9@gmail.com>