base: Lockscreen Date Fonts [1/2]
SDB: Call SliceView(DateView) Fonts with separate method. It looks neat and clean

Good Job @eldainosor. I was confused because of sliceview shit.
Bootleggers (@eldainosor) edit: Update the code to use the mKeyguardSlice instead of mDateView (deprecated due to google switching all the ls date section to slices) and make a function to apply the same typeface to every view on such SliceView

Signed-off-by: Aston-Martinn <advaithbhat9@gmail.com>
Change-Id: I5c41ad27aa82e0112d1873264f4214301f40874d