base: Try to fix upstream status bar+keyguard bug
It seems an upstream bug comes from `mLeaveOpenOnKeyguardHide` inside of
StatusBarStateControllerImpl.java, The bug seems to be that
`mLeaveOpenOnKeyguardHide` is left at `true` when the status bar is
closed, resulting in some weird behavior after the keyguard is hidden
(probably any code that reads that boolean field expects the status bar
to still be open: "leaveOpen").

Both `StatusBar.postQSRunnableDismissingKeyguard` and the lockscreen
notification expansion gesture code can set `mLeaveOpenOnKeyguardHide`
to true, leading to this bug. Able to reproduce the bug via the
notification gesture on the stock OS (sargo-rp1a.200720.009). See
https://github.com/GrapheneOS/platform_frameworks_base/pull/34#issuecomment-690777142
for more details. Was not able to reproduce this via the notification
gesture on an Android 10 (GrapheneOS QQ3A.200805.001.2020.08.07.01)
release.

How to reproduce for the tiles:
1. tapping any sensitive tile on the lockscreen,
2. pressing the back button to close the keyguard input,
3. collapsing the notification panel completely, and waiting for a few
   seconds,
4. unlocking the lock screen to hide keyguard.

Then, the panel bar no longer becomes expandable until you keep
trying.

This attempts to fix/workaround the bug by setting mLeaveOpenOnKeyguardHide
to be false when the status bar panel is closed.

Change-Id: Ic567a890c1147a41b9a486d97b1d3f9d3a4ff446
Signed-off-by: Aston-Martinn <advaithbhat9@gmail.com>